Support use of encoding.TextUnmarshaler with anchors
`encoding.TextUnmarshaler` is intended to be "the interface
implemented by an object that can unmarshal a textual representation
of itself".

By providing the text without the anchor, this allows common
types (e.g. `net.IP`, `big.Rat`) to be used as target values for
decoding with anchors.

Without this change, the text "&a 1" would be used instead of "1":

	---
	a: &a 1
	b: *a