scanner: Find quoted token from context
To resolve an issue with quoted map keys not correctly parsing, this PR
changes how it progresses after scanning a quote token.

The old behaviour was:
1. scanner reaches a :
2. pull the characters out of the current scan context buffer and turn it into the map key token
3. set the previous indent column to the beginning of the buffer
So next time it gets to another key in the map, it'll check whether the indent was the same.

When scanning a quote:
1. Scan from the beginning to the end of the quote sequence
2. Add a token, send back to the lexer, and reset the scan context
In this scenario, it reaches the :, but nothing is in the scan context buffer because it just went ahead and added the quote token. The previous indent column doesn't get reset, so when it gets to the next map key it thinks the indent increased, which is invalid yaml.

The new behaviour is essentially to not reset the context after scanning a quote, and when reaching a : check the context tokens, and if it's a quote set the previous indent column to that