scanner: resolve off-by-one scanning single quote
When scanning an escaped single quote `''` there was an off by one error
because while the local `idx` variable is increased, the context's `idx`
is not. This leads to the progress returned being off by one, so the
scanner gets stuck in the `ctx.next()` loop forever, unable to progress.