Do quote YAML 1.1 bools for compatibility with other legacy parsers
Quote YAML 1.1 bools like "on", "off", and so on at the encoding time so that the YAML data produced by the `goccy/go-yaml`'s can be fed into other legacy YAML parsers that supports YAML up to 1.1.

I do know go-yaml is intended to support YAML 1.2 only. However, this change does not break that. This enhances compatibility with legacy parsers without quitting to be a YAML 1.2 only library, because it's fine to encode a string literal `on` as `on` or `"on"`.

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>