Allow customizing the error message
This package produces really nice error messages that helpfully point to
a section of YAML that caused an error. When there's an error, the
package has technical knowledge of what the problem is. The calling
program may have more practical knowledge of what the problem is,
allowing it to present a helpful message to the user. In this case, it
can be useful to allow that calling program to modify the message, but
keep the portion of the error that points at the YAML source. Adding
this interface allows it to do so in a consistent way that's agnostic of
the exact underlying error type.

```go
func (o *SomeType) UnmarshalYAML(unmarshal func(interface{}) error) error {
  ...
  if err := unmarshal(&var); err != nil {
    if e, ok := err.(yaml.ErrorWithSource); ok {
      return e.WithMessage("Options for this field are ...")
    }
    return err
  }
  ...
}
```