add an encoding option to force string value quoting
The current encoder does not quote strings (with very few exceptions):
```yaml
foo: a b c
bar: xzy
baz: "true"
```

While this is perfectly valid YAML, it can be inconsistent and hard to read. A
common convention is to always quote string values. With the previous example:
```yaml
foo: "a b c"
bar: "xzy"
baz: "true"
```

Quoting keys is unnecessary, and this presentation is both consistent and easy
to read.

This patch adds an encoding option that always quotes string values. Hopefully I
understood the calls to encodeString correctly! Of course it would be possible
to add a quoteStringKeys option later if someone wants quoted strings
everywhere.

Context: we are using YAML to obtain a readable representation of API data
objects for debugging, and having consistent quoting everywhere helps.