Fix #687: zero-length strings from TextMarshaller are improperly handled
This issue is a combination of two: the first is that a naive zero length
string creates a nil valued node, which is in turned rendered as the
literal <nil>.

The second is that an empty string from a TextMarshaller is converted into
a nil valued node, rather then an empty string - which in YAML requires
explicit quotes to be included.

The origin of this problem is that output from TextMarshaller objects is
parsed as YAML, rather then as a string. This means amongst other issues that
a TextMarshaller which outputs a string which looks like valid YAML or JSON
will produce entirely the wrong content for re-marshalling the same object.

This commit fixes the second problem.

Actual fix