Add reproducer test for handling zero-length []byte
MarshalYAML functions which return bytes can plausibly produce 0-length outputs,
for example when trying to represent YAML `null` which allowably is show by the
empty string.

However `nil` parsed nodes from this documents lead to the literal value `<nil>`
being rendered into the returned document, which is then re-read as a string
and thus breaks round-tripped parsing.

This adds a test for the `<nil>` rendering problem s applied to the
ByteMarshaller functions which should implement `MarshalYAML` which in turn
should return a byte-array of valid YAML - of which a 0-length string is valid.