Change the TextMarshaller compliance test
The TextMarshaller test currently assumes that the output of TextMarshaller
is parsed as YAML, and in turn can then be turned into a YAML number in this
test.

This is incorrect if the value were to be round-tripped with the expectation
of preserving types, and thus this test must change.

This may or may not be a breaking change for some use cases, but success here
is dependent on incorrect behavior when dealing with more complicated
TextMarshaller output.