Add decode option that allows specific field prefixes (#795)
This would be paired with DisallowUnknownField where you generally want
to make sure the yaml does not have unknown fields, but allow for cases
where the yaml has some extra stuff in it, e.g. for custom metadata or
using fields as variables for anchors.

In our specific case we allow `x-*` fields for these cases.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>